Remove the second axios
