+++ src/action/Progress.js
@@ -1,5 +1,8 @@
+/* eslint-disable no-console */
 import axios from 'axios';
-import { FETCH_PROGRESS_REQUEST, FETCH_PROGRESS_SUCCESS, FETCH_PROGRESS_FAILURE } from '../store/type';
+import {
+  FETCH_PROGRESS_REQUEST, FETCH_PROGRESS_SUCCESS, FETCH_PROGRESS_FAILURE, ADD_PROGRESS,
+} from '../store/type';
 
 const fetchProgresRequest = () => ({
   type: FETCH_PROGRESS_REQUEST,
@@ -16,6 +19,11 @@ const fetchProgressFailure = (error) => ({
   payload: error,
 
 });
+export const addProgres = (progObj) => ({
+  type: ADD_PROGRESS,
+  payload: { progObj },
+
+});
 
 const fetchProgress = (id = 1) => (dispatch) => {
   dispatch(fetchProgresRequest);
@@ -29,7 +37,18 @@ const fetchProgress = (id = 1) => (dispatch) => {
       dispatch(fetchProgressFailure(errorMsg));
     });
 };
+const addProgress = (prog) => (dispatch) => {
+  axios.post(`https://intense-spire-98414.herokuapp.com/measurments/${prog.id}/results?result=${prog.result}`)
+    .then((response) => {
+      const progress = response.data;
+      dispatch(addProgres(progress));
+    })
+    .catch((error) => {
+      const errorMsg = error.message;
+      console.log(errorMsg);
+    });
+};
 
 export {
-  fetchProgressFailure, fetchProgresSuccess, fetchProgresRequest, fetchProgress,
+  fetchProgressFailure, fetchProgresSuccess, fetchProgresRequest, fetchProgress, addProgress,
 };
